refactor: runners on same folder, introducing utils.js
