+++ bench/scripts/load-test.js
