+++ bench/scripts/smoke-test.js
