+++ bench/scripts/spike-test.js
